Consolidated ModelChoiceField tests.
